requirements.txt
asgiref==3.3.4
astroid==2.5.6
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-braces==1.14.0
django-cors-headers==3.7.0
django-jazzmin==2.4.7
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
docutils==0.17.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==0.8
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==21.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.2
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
wrapt==1.12.1
